# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API pour la gestion des BAES",
        "title": "BAES API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Local Flask server"
        }
    ],
    "paths": {
...
        "/sites/": {
            "get": {
                "description": "Récupère la liste de tous les sites.",
                "responses": {
                    "200": {
                        "description": "Liste des sites.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "example": 1,
                                                "type": "integer"
                                            },
                                            "name": {
                                                "example": "Site 1",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne."
                    }
                },
                "tags": [
                    "Site CRUD"
                ]
            },
            "post": {
                "description": "Crée un nouveau site.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "example": "Site 1",
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ],
                                "type": "object"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Site créé avec succès.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "example": 1,
                                            "type": "integer"
                                        },
                                        "name": {
                                            "example": "Site 1",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mauvaise requête."
                    }
                },
                "tags": [
                    "Site CRUD"
                ]
            }
        },
...
        "/sites/{site_id}": {
            "delete": {
                "description": "Supprime un site par son ID et tous les éléments associés (bâtiments, étages, BAES, erreurs, cartes) en cascade. Les liaisons user-site-role sont conservées avec site_id=NULL pour maintenir les relations user-role.",
                "parameters": [
                    {
                        "description": "ID du site à supprimer",
                        "in": "path",
                        "name": "site_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Site supprimé avec succès.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "properties": {
                                        "baes_deleted": {
                                            "example": 10,
                                            "type": "integer"
                                        },
                                        "batiments_deleted": {
                                            "example": 2,
                                            "type": "integer"
                                        },
                                        "cartes_deleted": {
                                            "example": 3,
                                            "type": "integer"
                                        },
                                        "erreurs_deleted": {
                                            "example": 15,
                                            "type": "integer"
                                        },
                                        "etages_deleted": {
                                            "example": 5,
                                            "type": "integer"
                                        },
                                        "message": {
                                            "example": "Site supprimé avec succès",
                                            "type": "string"
                                        },
                                        "user_site_roles_preserved": {
                                            "example": 4,
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Site non trouvé."
                    }
                },
                "tags": [
                    "Site CRUD"
                ]
            },
            "get": {
                "description": "Récupère un site par son ID.",
                "parameters": [
                    {
                        "description": "ID du site à récupérer",
                        "in": "path",
                        "name": "site_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du site.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "example": 1,
                                            "type": "integer"
                                        },
                                        "name": {
                                            "example": "Site 1",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Site non trouvé."
                    }
                },
                "tags": [
                    "Site CRUD"
                ]
            },
            "put": {
                "description": "Met à jour un site existant.",
                "parameters": [
                    {
                        "description": "ID du site à mettre à jour",
                        "in": "path",
                        "name": "site_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Body2"
                },
                "responses": {
                    "200": {
                        "description": "Site mis à jour avec succès.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "example": 1,
                                            "type": "integer"
                                        },
                                        "name": {
                                            "example": "Site mis à jour",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Site non trouvé."
                    }
                },
                "tags": [
                    "Site CRUD"
                ]
            }
        },
...
        "/user_site_role/user/{user_id}/site/{site_id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "ID de l'utilisateur",
                        "in": "path",
                        "name": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "ID du site",
                        "in": "path",
                        "name": "site_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissions supprimées",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "remaining_permissions": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    }
                },
                "summary": "Supprime toutes les permissions d'un utilisateur pour un site donné",
                "tags": [
                    "UserSiteRole"
                ]
            }
        },
...
        "/users/sites/{user_id}/sites/{site_id}": {
            "delete": {
                "description": "Dissocie un site d'un utilisateur.",
                "parameters": [
                    {
                        "description": "ID de l'utilisateur",
                        "in": "path",
                        "name": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "ID du site à dissocier",
                        "in": "path",
                        "name": "site_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Site dissocié avec succès.",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Site dissocié de l'utilisateur.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur ou site non trouvé, ou site non associé."
                    },
                    "500": {
                        "description": "Erreur interne."
                    }
                },
                "tags": [
                    "User-Site Relations"
                ]
            }
        },
...
    },
...
    "components": {
        "requestBodies": {
...
            "Body2": {
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "example": "Site mis à jour",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "required": true
            }
        },
...
    }
}